Drop support for hotshot
For some reason coverage.py started to ignore my `# pragma: PY2` and
started marking the code for HotShotFuncProfile.__call__ and all of the
following HotShotFuncProfile and HotShotFuncCoverage methods as not
executed.  This is likely a bug in coverage.py, but all it did was
remind me to take out the trash out of the codebase.